This is not currently an evolving list
This commit changes the README to indicate that the current `live.md` is a snapshot from March 2014 rather than an evolving list. I know @timblair has plans to update this and keep it current, but I'm about to link to it in a blog post so would like to make it clear what the current state is.